02-Working-with-libraries/02-libraries-for-development/06-REST-and-auth/requirements.txt
fastapi
fastapi-code-generator
uvicorn
PyJWT
cryptography
bcrypt
fastapi_keycloak_middleware